debug v2

+++ src/geometry.py
@@ -55,15 +55,16 @@ def create_square2(square):
         [x_center - size_half, y_center + size_half],
         [x_center + size_half, y_center + size_half],
         [x_center + size_half, y_center - size_half]
-    ], dtype=torch.float32)
+    ])
 
     # Rotate the vertices according to the given rotation
-    rotation = torch.tensor(rotation, dtype=torch.float32)
+    rotation = torch.tensor(rotation)
     rotation_matrix = torch.tensor([
         [torch.cos(rotation), -torch.sin(rotation)],
         [torch.sin(rotation), torch.cos(rotation)]
     ])
 
+    # make all values from float->double
     rotated_vertices = torch.mm(vertices - torch.tensor([x_center, y_center]), rotation_matrix.t()) + torch.tensor([x_center, y_center])
 
     return rotated_vertices
